Move classes to DataType namespace.
